# AlamiaAccounts-Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AlamiaAccounts API",
        "description": "API testing",
        "version": "1.0.0"
    },
    "paths": {
...
        "/voucher-types": {
            "get": {
                "tags": [
                    "Voucher Types"
                ],
                "summary": "Get all custom voucher types",
                "operationId": "f30ab0fcfcc5000c5716325e6eeb65cd",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/voucher-types/{id}": {
            "get": {
                "tags": [
                    "Voucher Types"
                ],
                "summary": "Get a specific voucher type",
                "operationId": "fc5be1a27a24949652868ec037e0bbc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Voucher type not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/trial-balance": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Trial Balance report",
                "operationId": "9df9c5ab7f15f22b2654f8aca718e6c6",
                "parameters": [
                    {
                        "name": "as_of_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/profit-loss": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Profit and Loss report",
                "operationId": "8be49de6901b14e52e093123e1b21c2f",
                "parameters": [
                    {
                        "name": "from_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/balance-sheet": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Balance Sheet report",
                "operationId": "0b5db21d750b8e7ccde0e02718acae44",
                "parameters": [
                    {
                        "name": "as_of_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}